feat: add ability to add data point directly
Previously, it was assumed that data would come in as a JSON
object and logic was implemented to traverse the object and add
each leaf data point in.

This would require additional processing from the client component
however, to convert a jsonpath to an actual object. Now, redact-store
can take in a plain data object and put it in the DB.